Pull status out into a separate page
