+++ jobserver/settings.py
@@ -80,7 +80,6 @@ TEMPLATES = [
                 "django.contrib.auth.context_processors.auth",
                 "django.contrib.messages.context_processors.messages",
                 "jobserver.context_processors.nav",
-                "jobserver.context_processors.site_stats",
             ],
         },
     },
