+++ jobserver/urls.py
@@ -27,6 +27,7 @@ from .views import (
     JobRequestList,
     JobRequestZombify,
     JobZombify,
+    Status,
     WorkspaceCreate,
     WorkspaceDetail,
     WorkspaceLog,
@@ -53,6 +54,7 @@ urlpatterns = [
     path("jobs/<pk>/", JobDetail.as_view(), name="job-detail"),
     path("jobs/<pk>/zombify/", JobZombify.as_view(), name="job-zombify"),
     path("logout/", LogoutView.as_view(), name="logout"),
+    path("status/", Status.as_view(), name="status"),
     path("workspaces/", RedirectView.as_view(url="/")),
     path("workspaces/new/", WorkspaceCreate.as_view(), name="workspace-create"),
     path("__debug__/", include(debug_toolbar.urls)),
